% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detects.R
\name{summarizeValues}
\alias{summarizeValues}
\title{Summarize NIBRS data to tally percent of given values for a variable}
\usage{
summarizeValues(extractDfList, variableValueList)
}
\arguments{
\item{extractDfList}{a list of NIBRS data frames
At a minimum, this list must include \code{FullIncidentView}, \code{State} and \code{Agency} data frames.}

\item{variableValueList}{A list structure configuring the summarization. The structure
should contain the following names (keys) and values:
\itemize{
\item \code{aggregateBy}: a character vector containing the names of the columns in \code{FullIncidentView} that should be distinctly counted
\item \code{stateRegex}: a regular expression representing the two-letter state codes to include in the analysis
\item \code{yearRegex}: a regular expression representing the years to include in the analysis
\item 1 or more sub-lists, where the name (key) of the sublist in \code{variableValueList} is the name of the column to summarize, and the
elements of the list have the following names (keys) and values:
\itemize{
\item \code{values}: the values of the column to count as "anomalies" (i.e., to count in the numerator of the tally)
\item \code{label}: a label to use as the column name for the resulting tally of this column
}
}}
}
\value{
a data frame with the following columns:
\itemize{
\item \code{AgencyName}: The agency name
\item \code{Records}: The number of distinct-counted values of the \code{aggregateBy} columns for that agency
\item A column for each of the variables identified in \code{variableValueList}, containing the percentage of records for each agency
for which the value of the variable matches the \code{values}
}
}
\description{
This function summarizes data in the NIBRS \code{FullIncidentView} materialized view, by performing a distinct-count tally
of records for each agency, and calculating a percentage of the tallied records that match a defined list of values. Primarily
this is useful for anomaly detection, but can be used for tallying any range of values.
}
